# BikeWorldServer/api-docs.yaml
openapi: 3.0.0
info:
  title: bike-world API
  version: '2.0'
servers:
  - url: https://bike-world-ids.herokuapp.com/
paths:
...
  /api/v2/itineraries/{id}:
    put:
      summary: Update an itinerary.
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the rental point to update.
          schema:
            type: string
            example: '627fb874a991e0a164e063ac'
        - name: x-access-token
          in: header
          required: true
          description: The jwt of the session.
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                addressStarting:
                  type: string
                  example: 'Trento'
                latS: 
                  type: number
                  example: 45.4654219
                lngS: 
                  type: number
                  example: 10.1859243
                description:
                  type: string
                  example: "Giro molto bello"
                difficulty:
                  type: string
                  example: "Principiante"
                length:
                  type: number
                  example: 8.2
      responses:
        '200':
          description: Itinerary updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Itinerary info updated!"
        '404':
          description: Itinerary not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Itinerary not found"
        '401':
          description: Unauthorized. No token provided.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'No token provided.'
        '403':
          description: Forbidden. Failed to authenticate token.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Failed to authenticate token.'
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
  /api/v2/itineraries/difficulty:
    get:
      summary: Get itineraries by difficulty.
      parameters:
        - name: type
          in: query
          required: true
          description: The difficulty of the itinerary to return.
          schema:
            type: string
            example: 'Principiante'
      responses:
        '200':
          description: List of the itineraries with a specific difficulty.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: 'Giro dei laghi'
                    addressStarting:
                      type: string
                      example: 'Trento'
                    latS: 
                      type: number
                      example: 45.4654219
                    lngS: 
                      type: number
                      example: 10.1859243
                    description:
                      type: string
                      example: "Giro molto bello"
                    difficulty:
                      type: string
                      example: "Principiante"
                    length:
                      type: number
                      example: 8.2
                    self:
                      type: string
                      example: "/api/v2/itineraries/627fb874a991e0a164e063ac"   
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
  /api/v2/itineraries/zone:
    get:
      summary: Get itineraries by zone.
      parameters:
        - name: latitude
          in: query
          required: true
          description: The latitude of the itineraries to return.
          schema:
            type: number
            example: 45.4654219
        - name: longitude
          in: query
          required: true
          description: The longitude of the itineraries to return.
          schema:
            type: number
            example: 10.1859243
      responses:
        '200':
          description: List of the itineraries in a specific zone.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: 'Giro dei laghi'
                    addressStarting:
                      type: string
                      example: 'Trento'
                    latS: 
                      type: number
                      example: 45.4654219
                    lngS: 
                      type: number
                      example: 10.1859243
                    description:
                      type: string
                      example: "Giro molto bello"
                    difficulty:
                      type: string
                      example: "Principiante"
                    length:
                      type: number
                      example: 8.2
                    self:
                      type: string
                      example: "/api/v2/itineraries/627fb874a991e0a164e063ac"   
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
  /api/v2/itineraries/length:
    get:
      summary: Get itineraries based on the length.
      parameters:
        - name: minLength
          in: query
          required: true
          description: The minLength used to check the length of the itinerary.
          schema:
            type: number
            example: 5
        - name: maxLength
          in: query
          required: true
          description: The maxLength used to check the length of the itinerary.
          schema:
            type: number
            example: 50
      responses:
        '200':
          description: List of the itineraries that are in a specific range of length.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    name:
                      type: string
                      example: 'Giro dei laghi'
                    addressStarting:
                      type: string
                      example: 'Trento'
                    latS: 
                      type: number
                      example: 45.4654219
                    lngS: 
                      type: number
                      example: 10.1859243
                    description:
                      type: string
                      example: "Giro molto bello"
                    difficulty:
                      type: string
                      example: "Principiante"
                    length:
                      type: number
                      example: 8.2
                    self:
                      type: string
                      example: "/api/v2/itineraries/627fb874a991e0a164e063ac"   
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'

  ###################
  # route damages
  ###################
...